linear/requirements.txt
mysql-connector-python==8.0.33
numpy==1.24.3
protobuf==3.20.3
PuLP==2.7.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
